requirements.txt
# Chess game requirements
python-chess==1.999.0
openai==1.12.0
